requirements.txt
flask
flask-socketio
python-socketio
elevenlabs
python-dotenv
openai
requests
arrow==1.3.0
dateparser==1.2.0
pytz
pyaudio
psutil
